+++ src/main/java/amsSdkDemo/Demo.java
@@ -9,6 +9,12 @@ import java.util.Map;
 
 import com.alipay.ams.AMS;
 import com.alipay.ams.AMSClient;
+import com.alipay.ams.callbacks.PaymentCancelCallback;
+import com.alipay.ams.callbacks.PaymentContextCallback;
+import com.alipay.ams.callbacks.PaymentInquiryCallback;
+import com.alipay.ams.callbacks.PaymentRefundCallback;
+import com.alipay.ams.callbacks.UserPresentedCodePaymentCallback;
+import com.alipay.ams.callbacks.impl.InMemoryPaymentContextCallback;
 import com.alipay.ams.cfg.AMSSettings;
 import com.alipay.ams.domain.Amount;
 import com.alipay.ams.domain.AuthNotifyModel;
@@ -19,12 +25,6 @@ import com.alipay.ams.domain.Order;
 import com.alipay.ams.domain.PaymentResultModel;
 import com.alipay.ams.domain.ResponseResult;
 import com.alipay.ams.domain.Store;
-import com.alipay.ams.domain.callbacks.PaymentCancelCallback;
-import com.alipay.ams.domain.callbacks.PaymentContextCallback;
-import com.alipay.ams.domain.callbacks.PaymentInquiryCallback;
-import com.alipay.ams.domain.callbacks.PaymentRefundCallback;
-import com.alipay.ams.domain.callbacks.UserPresentedCodePaymentCallback;
-import com.alipay.ams.domain.callbacks.impl.InMemoryPaymentContextCallback;
 import com.alipay.ams.domain.requests.PaymentCancelRequest;
 import com.alipay.ams.domain.requests.PaymentInquiryRequest;
 import com.alipay.ams.domain.requests.PaymentRefundRequest;
