+++ src/main/java/com/alipay/ams/util/LockUtil.java
@@ -6,7 +6,7 @@ package com.alipay.ams.util;
 
 import java.util.concurrent.TimeUnit;
 
-import com.alipay.ams.domain.callbacks.PaymentContextCallback;
+import com.alipay.ams.callbacks.PaymentContextCallback;
 
 /**
  *
